+++ src/memory_bus.rs
@@ -20,7 +20,7 @@ impl MemoryBus {
         self.memory.read_rom(f)
     }
 
-    pub fn read_byte(&self, address: usize) -> u8 {
+    pub fn read_byte(&self, address: u16) -> u8 {
         match address {
             0..=0x7FFE => self.memory.read_byte(address),
             0x8000..=0x9FFF => self.gpu.read_byte(address),
@@ -30,7 +30,7 @@ impl MemoryBus {
     }
 
     // TODO return errors?
-    pub fn write_byte(&mut self, address: usize, val: u8) {
+    pub fn write_byte(&mut self, address: u16, val: u8) {
         match address {
             0..=0x7FFE => self.memory.write_byte(address, val),
             0x8000..=0x9FFF => self.gpu.write_byte(address, val),
